% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyze_coxph}
\alias{analyze_coxph}
\title{Wrapper for Cox Proportional Hazards Modeling}
\usage{
analyze_coxph(data, formula, submodel_var, submodel_value)
}
\arguments{
\item{data}{Input data.frame}

\item{formula}{A character vector specifying the model formula}

\item{submodel_var}{Optional: A character vector specifying variable within data which is used for subgroup analyses. Defaults to NULL.}

\item{submodel_value}{Optional: A character vector specifying the values of the variable that are used for subgroup analyses. Defaults to NULL.}
}
\value{
An object of class cox model (check and correct this!)
}
\description{
Wrapper for Cox Proportional Hazards Modeling
}
