% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survtable.R
\name{create_survtable}
\alias{create_survtable}
\title{Create survival model specification tables}
\usage{
create_survtable(
  exposure_vars,
  outcome_vars,
  data_name,
  time_var,
  covariates = default_covariates,
  model_type = "time_invariant",
  submodel_var = NULL,
  submodel_values = NULL
)
}
\arguments{
\item{exposure_vars, time_var, outcome_vars, data_name}{Input string vectors specifying
variables for exposure_vars of primary interest, follow-up time_var, and outcome_vars
in survival analysis as well as survival model data_name. These are passed to
tidyr::crossing() to produce all possible combinations of the 4 vectors.
Note! \code{exposure_vars}, \code{time_var}, and \code{outcome_vars} refer to variable names in the
target data_name frame whereas \code{data_name} refers to a data_name.frame in the R
environment.}

\item{covariates}{Input string with length 1 specifying other variables that
are adjusted for. Defaults to default_covariates, an external vector.}

\item{model_type}{Input string specifying whether the model type is
time_var-invariant ("ti" or "time_var_invariant) or time_var-varying ("tv" or
"time_var_varying"). Defaults to "time_var_invariant"}

\item{submodel_var}{Optional. Input string specifying the variable in study
data_name that specifies subgroups. Defaults to NULL.}

\item{submodel_values}{Optional. Input string specifying the values of
submodel_var that define subgroups. Defaults to NULL.}
}
\value{
Hopefully one day an obejct with class \code{survtable}. Currently a
tibble that includes a model data_name formula for survival analysis as well as
individual variables that were used to produce them.
}
\description{
Builds a data.frame that includes
}
\examples{
survtable1 <- create_survtable(
  exposure_vars = c("exposure_2cat", "exposure_continuous"),
  outcome_vars = c("outcome1", "outcome2"),
  covariates = "age + sex + hla",
  time_var = "cens_time",
  data_name = "example_ti"
)

}
