% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ti.R
\docType{data}
\name{example_ti}
\alias{example_ti}
\title{Example data for time-invariant survival analysis}
\format{
\subsection{\code{example_ti}}{

A data frame with 12,000 rows and 9 columns:
\describe{
\item{id}{A subject-specific identifier}
\item{cens_time}{Time-to-event}
\item{outcome1, outcome2}{Examples of outcomes}
\item{age}{Example of a covariate}
\item{hla}{Example of a covariate}
\item{exposure_2cat}{Example exposure (categorical, 2 categories)}
\item{exposure_continuous}{Example exposure (continuous)}
\item{sex}{Sex}
...
}
}
}
\source{
Simulated data
}
\usage{
example_ti
}
\description{
Simulated dataset for survival analyses with a time-invariant exposure of interest
}
\keyword{datasets}
