% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tv.R
\docType{data}
\name{example_tv}
\alias{example_tv}
\title{Example data for time-variant survival analysis}
\format{
\subsection{\code{example_tv}}{

A data frame with 8,000 rows and 9 columns:
\describe{
\item{id}{A subject-specific identifier}
\item{tstart, tstop}{Start and stop times for each interval}
\item{event}{A binary indicator of whetehr the event occurred}
\item{exposure_continuous, exposure_2cat}{Examples of time-varying exposures/covariates}
...
}
}
}
\source{
Simulated data
}
\usage{
example_tv
}
\description{
Simulated dataset for survival analyses with a time-varying exposure of interest
}
\keyword{datasets}
