% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_resids.R
\name{export_resids}
\alias{export_resids}
\title{Export residuals}
\usage{
export_resids(
  caught_models,
  model_list,
  path = default_path,
  height = default_height,
  width = default_width
)
}
\arguments{
\item{caught_models}{A data.frame, output from catch_nonph()}

\item{model_list}{A list of Cox proportional hazards models, created by model_survtable()}

\item{path}{filepath to output folder. Defaults to an external vector \emph{default_path}}

\item{height}{Height of the output png-file. Defaults to external vector \emph{default_height}}

\item{width}{Width of the output png-file. Defaults to external vector \emph{default_weight}}
}
\value{
Prints schoenfeld residuals using ggcoxdiagnostics from 'survminer' packge
}
\description{
Export residuals
}
