% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefs.R
\name{get_coefs}
\alias{get_coefs}
\title{Get Coefficients of Interest in Models}
\usage{
get_coefs(model_list, select_terms)
}
\arguments{
\item{model_list}{A list of Cox Proportional Hazards Models.}

\item{select_terms}{A character vector of terms (typically the same as exposure_vars in create_survtable).}
}
\value{
A data.frame that includes model coefficients.
}
\description{
Get Coefficients of Interest in Models
}
\examples{
create_survtable(
  exposure_vars = c("exposure_2cat", "exposure_continuous"),
  outcome_vars = c("outcome1", "outcome2"),
  covariates = "age + sex + hla",
  time_var = "cens_time",
  data_name = "example_ti") |>
  model_survtable() |>
  get_coefs(c("exposure_2cat", "exposure_continuous"))

}
