% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_meta.R
\name{get_model_meta}
\alias{get_model_meta}
\title{Get Model Metadata}
\usage{
get_model_meta(model_list)
}
\arguments{
\item{model_list}{List of Cox Proportional Hazard Models}
}
\value{
Data frame describing the model formula, the total number of observations in analyses, number of events, number of missing observations.
}
\description{
Get Model Metadata
}
\examples{

create_survtable(
  exposure_vars = c("exposure_2cat", "exposure_continuous"),
  outcome_vars = c("outcome1", "outcome2"),
  covariates = "age + sex + hla",
  time_var = "cens_time",
  data_name = "example_ti") |>
  model_survtable() |>
  get_model_meta()

}
