% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_coefs.R
\name{graph_coefs}
\alias{graph_coefs}
\title{Draw Forest Plot}
\usage{
graph_coefs(data, title)
}
\arguments{
\item{data}{Model coefficients.}

\item{title}{Graph title.}
}
\value{
Forest plot (ggplot object)
}
\description{
Draw Forest Plot
}
\examples{
create_survtable(
  exposure_vars = c("exposure_2cat", "exposure_continuous"),
  outcome_vars = c("outcome1", "outcome2"),
  covariates = "age + sex + hla",
  time_var = "cens_time",
  data_name = "example_ti") |>
  model_survtable() |>
  get_coefs(c("exposure_2cat", "exposure_continuous")) |>
  graph_coefs(title = "**Your title here**")#'
}
