% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_resids.R
\name{graph_schoenfeld}
\alias{graph_schoenfeld}
\title{Graph Schoenfeld Residuals}
\usage{
graph_schoenfeld(
  fit,
  title = NULL,
  hline.col = "red",
  hline.size = 1,
  hline.alpha = 1,
  hline.yintercept = 0,
  hline.lty = "dashed",
  sline.col = "blue",
  sline.size = 1,
  sline.alpha = 0.3,
  sline.lty = "dashed",
  point.col = "black",
  point.size = 1,
  point.shape = 19,
  point.alpha = 1,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{fit}{Cox Proportional Hazards model fit produced by \emph{model_survtable()}}

\item{title, subtitle, caption}{optional formatting passed to ggplot2::labels()}

\item{hline.col, hline.size, hline.alpha, hline.yintercept, hline.lty}{optional formatting passed to ggplot2::geom_hline()}

\item{sline.col, sline.size, sline.alpha, sline.lty}{optional formatting passed to ggplot2::geom_smooth()}

\item{point.col, point.size, point.shape, point.alpha}{optional formatting passed to ggplot2::geom_point()}
}
\value{
ggplot2 object that contains schoenfeld residuals
}
\description{
Graph Schoenfeld Residuals
}
