% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_survtable.R
\name{model_survtable}
\alias{model_survtable}
\title{Model Survival Using Cox Proportional Hazards Models}
\usage{
model_survtable(survtable, cluster = NULL)
}
\arguments{
\item{survtable}{A data.frame containing \code{data_name}, \code{formula_str}, \code{submodel_var}, \code{submodel_value},
\code{outcome_var} and \code{exposure_var}}

\item{cluster}{Optional. Optional unquoted variable which clusters the observations,
for the purposes of a robust variance.  Defaults to NULL. See help(coxph)
for further details.}
}
\value{
A list of survival models with legnth of the number of rows in input data.frame.
}
\description{
Model Survival Using Cox Proportional Hazards Models
}
\examples{

create_survtable(
  exposure_vars = c("exposure_2cat", "exposure_continuous"),
  outcome_vars = c("outcome1", "outcome2"),
  covariates = "age + sex + hla",
  time_var = "cens_time",
  data_name = "example_ti") |>
model_survtable()

}
