% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_resids.R
\name{graph_resids}
\alias{graph_resids}
\title{Graph residuals}
\usage{
graph_resids(caught_models, model_list)
}
\arguments{
\item{caught_models}{A data.frame, output from catch_nonph()}

\item{model_list}{A list of Cox proportional hazards models, created by model_survtable()}
}
\value{
ggplot objects that assess the proportionality of hazards
}
\description{
Graph residuals
}
