% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_resids.R
\name{graph_resids}
\alias{graph_resids}
\title{Graph residuals}
\usage{
graph_resids(caught_models, model_list)
}
\arguments{
\item{caught_models}{A data.frame, output from catch_nonph()}

\item{model_list}{A list of Cox proportional hazards models, created by model_survtable()}
}
\value{
ggplot objects that assess the proportionality of hazards
}
\description{
Graph residuals
}
\examples{
library(dplyr)
models_2 <- create_survtable(exposure_vars = c("exposure_2cat", "exposure_continuous"),
outcome_vars = c("outcome1", "outcome2"),
covariates = "age + sex",
submodel_var = "hla",
submodel_values = c("Type 1", "Type 2", "Type 3"),
time_var = "cens_time",
data_name = "example_ti") \%>\%
model_survtable()

models_2 \%>\%
  catch_nonph() \%>\%
    graph_resids(models_2)
}
